requirements.txt
feedparser==6.0.11
python-telegram-bot==20.8
python-dotenv==1.0.1
schedule==1.2.0
requests==2.31.0